[rebase] origin/main에 buffer 기능 추가
