+++ ssd/virtual_ssd.py
@@ -4,7 +4,6 @@ from typing import Optional
 
 from logger.logger import Logger
 from ssd.buffer import Buffer
-from storage_device_interface import StorageDeviceInterface
 from ssd.storage_device_interface import StorageDeviceInterface
 import pandas as pd
 import os
@@ -15,6 +14,7 @@ INIT_VALUE = "0x00000000"
 class VirtualSSD(StorageDeviceInterface):
     def __init__(self) -> None:
         self.logger = Logger()
+        self.buffer = Buffer()
         self.nand_path = "nand.csv"
         self.result_path = "result.txt"
 
@@ -27,10 +27,10 @@ class VirtualSSD(StorageDeviceInterface):
         self.nand_df = pd.read_csv(self.nand_path)
 
     def write(self, address: int, data: str) -> None:
-        if buffer.get_size() == 10:
+        if self.buffer.get_size() == 10:
             self.flush()
         else:
-            buffer.add_cmd("W", address, data)
+            self.buffer.add_cmd("W", address, data)
 
         self.logger.print("Data has been successfully written to the SSD.")
 
@@ -39,8 +39,8 @@ class VirtualSSD(StorageDeviceInterface):
             os.remove(self.result_path)
 
         read_data = None
-        if buffer.get_size() > 0:
-            read_data = buffer.read_addressvalue_in_cmdlist(address)
+        if self.buffer.get_size() > 0:
+            read_data = self.buffer.read_addressvalue_in_cmdlist(address)
         if read_data is None:
             read_data = self.nand_df.loc[address, "Data"]
         elif read_data == "Erase":
@@ -54,15 +54,15 @@ class VirtualSSD(StorageDeviceInterface):
         self.logger.print("Data has been successfully read from the SSD.")
 
     def erase(self, address: int, size: int) -> None:
-        if buffer.get_size() == 10:
+        if self.buffer.get_size() == 10:
             self.flush()
         else:
-            buffer.add_cmd("E", address, size)
+            self.buffer.add_cmd("E", address, size)
 
         self.logger.print("SSD has been successfully erased.")
 
     def flush(self) -> None:
-        deque_buffer = deque(buffer.cmdlist)
+        deque_buffer = deque(self.buffer.cmdlist)
         while deque_buffer:
             old_command = deque_buffer.popleft()
             command1 = int(old_command[1])
@@ -80,31 +80,8 @@ class VirtualSSD(StorageDeviceInterface):
                         f"Flushing... : erase addr {command1} to addr {command1 + command2 - 1}"
                     )
         self.logger.print("SSD has been successfully flushed.")
-        buffer.flush()
+        self.buffer.flush()
         if os.path.exists(self.nand_path):
             os.remove(self.nand_path)
         self.nand_df["Data"].to_csv(self.nand_path, index_label="index")
         self.logger.print("SSD has been successfully flushed.")
-
-
-if __name__ == "__main__":
-    ssd = VirtualSSD()
-    buffer = Buffer()
-
-    cmd: Optional[str]
-    address: Optional[str]
-    value: Optional[str]
-
-    cmd, address, value = (sys.argv[1:4] + [None] * 3)[:3]
-
-    if cmd == "W":
-        if address is not None and value is not None:
-            ssd.write(int(address), value)
-    elif cmd == "R":
-        if address is not None:
-            ssd.read(int(address))
-    elif cmd == "E":
-        if address is not None and value is not None:
-            ssd.erase(int(address), int(value))
-    elif cmd == "F":
-        ssd.flush()
